examples/requirements.txt
gemini_json_agent
